Updated collision system to support moving platforms and fixed a bug in which the player's update method was called twice per frame.
The reason previous attempts to get the player to move along with a moving platform failed was that the player's update method being called twice in the GC's update method: once explicitely and once through the NPC update loop. Now that the player is only update donce per frame, the physics engine works smoothly. Every object now has a distinct friction, and if the player is in the air, they can't jump or move but no friction is applied so they maintain they original speed.
As every object now has its own friction, moving platforms will pull players to be at their velocity, letting players ride them.
Also added in an ice block with reduced friction as a demonstration.